tools/ocr_translations/requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.9.0
numpy>=1.26.0
requests>=2.31.0
PyYAML>=6.0
